+++ src/main/handlers/media/transcribe.ts
@@ -4,21 +4,15 @@ import { TranscriptionEngine, TranscriptionFunction } from './transcribeTypes';
 import deepspeechTranscribeFunction from './transcriptionEngines/deepspeechTranscribeFunction';
 import dummyTranscribeFunction from './transcriptionEngines/dummyTranscribeFunction';
 
-const getTranscriptionFunction = new Map<
+const getTranscriptionFunction: Record<
   TranscriptionEngine,
   TranscriptionFunction
->();
-
-// Add the enum - function mapping for a transcription engine here
-getTranscriptionFunction.set(TranscriptionEngine.VOSK, voskTranscribeFunction);
-getTranscriptionFunction.set(
-  TranscriptionEngine.DEEPSPEECH,
-  deepspeechTranscribeFunction
-);
-getTranscriptionFunction.set(
-  TranscriptionEngine.DUMMY,
-  dummyTranscribeFunction
-);
+> = {
+  // Add the enum - function mapping for a transcription engine here
+  [TranscriptionEngine.VOSK]: voskTranscribeFunction,
+  [TranscriptionEngine.DUMMY]: dummyTranscribeFunction,
+  [TranscriptionEngine.DEEPSPEECH]: deepspeechTranscribeFunction,
+};
 
 /**
  * Runs the transcription using the selected transcription engine
@@ -32,8 +26,7 @@ const transcribe = async (
   transcriptionEngine: TranscriptionEngine
 ) => {
   const transcriptionFunction =
-    getTranscriptionFunction.get(transcriptionEngine) ??
-    dummyTranscribeFunction;
+    getTranscriptionFunction[transcriptionEngine] ?? dummyTranscribeFunction;
   return transcriptionFunction(project);
 };
